Moved Rails initialization into a Railtie (fixes Rails 6 deprecation warnings, closes #206)
